02-workflow-orchestration/airflow_nyc/requirements/base.txt
apache-airflow-providers-google
pyarrow 
#remove this and place it toask or dag specific requirmenet with virtual env